Move all custom HiGlass tracks to separate directory
