+++ src/components/custom-tracks/README
@@ -1,0 +1,3 @@
+# Custom Tracks
+
+Directory for custom components directly extending HiGlassComponent to create custom HiGlass tracks.
